Fix visualization of hit and death
